refactor(server): split app startup
